+++ test/test_graph.jl
@@ -6,7 +6,7 @@
 
     # Create a mutable mask and set a bit
     mask = _MutableMask{1}()
-    _set_bit!(mask, UInt8(3))
+    _set_bit!(mask, 3)
 
     # Test _find_or_create: should create a new node
     node = _find_or_create(graph, mask)
@@ -21,31 +21,31 @@
 
     # Test _find_node: add and remove components
     start = graph.nodes[1]
-    add = (UInt8(5),)
+    add = (5,)
     remove = ()
 
     node3 = _find_node(graph, start, add, remove)
     @test node3 !== start
-    @test _get_bit(node3.mask, UInt8(5))
+    @test _get_bit(node3.mask, 5)
 
-    node4 = _find_node(graph, node3, (), (UInt8(5),))
+    node4 = _find_node(graph, node3, (), (5,))
     @test node4 === start
 
     # Test error on removing nonexistent component
     @test_throws(
         "ArgumentError: entity does not have component to remove",
-        _find_node(graph, start, (), (UInt8(7),))
+        _find_node(graph, start, (), (7,))
     )
 
     # Test error on adding duplicate component
     @test_throws(
         "ArgumentError: entity already has component to add, or it was added twice",
-        _find_node(graph, node3, (UInt8(5),), ())
+        _find_node(graph, node3, (5,), ())
     )
 
     # Test add and remove same
     @test_throws(
         "ArgumentError: component added and removed in the same exchange operation",
-        _find_node(graph, node3, (UInt8(5),), (UInt8(5),))
+        _find_node(graph, node3, (5,), (5,))
     )
 end
